backend/requirements.txt
# backend/requirements.txt
# Core Flask Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# ML Dependencies
tensorflow>=2.16.0
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
scikit-learn>=1.3.2
joblib>=1.3.2
numpy>=1.24.0,<2.0.0

# RAG & Vector Store Dependencies
groq>=0.4.0
faiss-cpu>=1.7.4

# File Processing
PyPDF2>=3.0.1
python-docx>=1.1.0

